initial commit yo
This is starting over as far as the original Cratchit is concerned. The
idea is to have everything admin-wise (and possibly everything speaker-
wise) handled through this app, using the same database as we use for
the standard CocoaConf app. We will probably use mostly standard Grails
scaffolding for this, although if we can figure out Twitter Bootstrap
at some point that might be a good thing to use instead for all I know.
(I don't know much.)

Marley was dead: to begin with. There is no doubt whatever about that.